+++ frontEnd_ontimize/src/app/main/personal-area/personal-area-detail/personal-area-detail.component.html
@@ -20,12 +20,9 @@ header-actions="" show-header="no"
   
   <o-table-column attr="name" title="{{'Nombre' | oTranslate }}" width="100px"></o-table-column>
   <o-table-column attr="document_type" title="{{'Tipo' | oTranslate }}" width="100px"></o-table-column>
-
-
-  <o-table-cell-renderer-action (onClick)="actionClick($event)" icon="download"></o-table-cell-renderer-action>
-
-
-
+  <o-table-column attr="action" width="64px">
+    <o-table-cell-renderer-action (onClick)="actionClick($event)" icon="download"></o-table-cell-renderer-action>
+  </o-table-column>
   
   </o-table>
