+++ frontEnd_ontimize/src/app/main/personal-area/personal-area-home/personal-area.component.ts
@@ -9,15 +9,12 @@ import { OntimizeService, OTableComponent } from 'ontimize-web-ngx';
 })
 export class PersonalAreaComponent implements OnInit {
 
-  protected productService: OntimizeService;
-
 
   public isAdmin: boolean;
   private myRoleService: OntimizeService;
   public selectedDocument: any;
 
   @ViewChild('table', {static: false }) public tableDocuments: OTableComponent;
-
 
   constructor(
 
@@ -26,13 +23,9 @@ export class PersonalAreaComponent implements OnInit {
     public injector: Injector
   ) {
     this.myRoleService = this.injector.get(OntimizeService);
-    this.productService = this.injector.get(OntimizeService);
-
   }
 
   ngOnInit() {
-    const conf = this.productService.getDefaultServiceConfiguration('products');
-    this.productService.configureService(conf);
     this.configureUserRoleService();
     this.myRoleService.query({}, ["rolename"], "myRole").subscribe(
       response => {
@@ -71,21 +64,5 @@ export class PersonalAreaComponent implements OnInit {
 
   }
 
-
-  actionClick(event){
-    this.productService.query({id:event.id}, ['name','base64'], 'fileContent').subscribe(res => {
-      if (res.data && res.data.length) {
-        let filename = res.data[0].name;
-        let base64 = res.data[0].base64;
-        const src = `data:text/csv;base64,${base64}`;
-        const link = document.createElement("a");
-        link.href = src;
-        link.download = filename;
-        link.click();
-        link.remove();
-      }
-    });
-    
-  }
 
 }
